Same classloader record metrics
